requirements.txt
fastapi~=0.97.0
uvicorn[standard]
pyannote.audio
https://github.com/pyannote/pyannote-audio/archive/refs/heads/develop.zip
openai